fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fixup! fix(hitl): rename respondants as assignees and user HITLUser to align the assignee, responded_by interface
