+++ airflow-core/tests/unit/api_fastapi/core_api/routes/public/test_hitl.py
@@ -31,7 +31,7 @@ from sqlalchemy.orm import Session
 from airflow._shared.timezones.timezone import utcnow
 from airflow.models.hitl import HITLDetail
 from airflow.models.log import Log
-from airflow.sdk.types import HITLUser
+from airflow.sdk.execution_time.hitl import HITLUser
 from airflow.utils.state import TaskInstanceState
 
 if TYPE_CHECKING:
